feat: move pages to separate directory
